idm-service/requirements.txt
grpcio
grpcio-tools
SQLAlchemy
pymysql
python-dotenv
pyjwt
cryptography